Local, Server & Manager unified binary ssservice
- fixes #698
- Refactored shadowsocks-rust crate structure. Common code are put into
a separated library
