+++ crates/shadowsocks-service/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Shadowsocks Contributors"]
 description = "shadowsocks is a fast tunnel proxy that helps you bypass firewalls."
 repository = "https://github.com/shadowsocks/shadowsocks-rust"
 readme = "README.md"
-documentation = "https://docs.rs/shadowsocks"
+documentation = "https://docs.rs/shadowsocks-service"
 keywords = ["shadowsocks", "proxy", "socks", "socks5", "firewall"]
 license = "MIT"
 edition = "2018"
